+++ src/connectedk8s/azext_connectedk8s/_troubleshootutils.py
@@ -24,7 +24,7 @@ from azext_connectedk8s._utils import get_utctimestring
 if TYPE_CHECKING:
     from kubernetes.client import AppsV1Api, BatchV1Api, CoreV1Api
 
-    from .vendored_sdks.preview_2024_07_01.models import (
+    from .vendored_sdks.preview_2025_08_01.models import (
         ConnectedCluster,
     )
 
@@ -1000,7 +1000,7 @@ def check_agent_version(
             return consts.Diagnostic_Check_Incomplete
 
         # To get user agent version and the latest agent version
-        user_agent_version = connected_cluster.agent_version  # type: ignore[unreachable]
+        user_agent_version = connected_cluster.agent_version
         current_user_version = user_agent_version.split(".")
         latest_agent_version = azure_arc_agent_version.split(".")
         # Comparing if the user version is compatible or not